+++ lib/rover_lib2/src/rover_lib2/actuators/actuator.hpp
@@ -3,58 +3,64 @@
 
 #include <rover_lib2/rover_object.hpp>
 
-template<typename Impl_T>
-class Actuator : public RoverObject<Actuator<Impl_T>>
+#include <optional>
+
+namespace Actuators
 {
-  private:
-    friend Impl_T;
-    Actuator() = default;
-
-  public:
-    void _init(void)
+    template<typename Impl_T>
+    class Actuator : public RoverObject<Actuator<Impl_T>>
     {
-        static_cast<Impl_T*>(this)->__init();
-    }
+      private:
+        friend Impl_T;
+        Actuator() = default;
 
-    void _update(void)
-    {
-        static_cast<Impl_T*>(this)->__update();
-    }
+      public:
+        void _init(void)
+        {
+            static_cast<Impl_T*>(this)->__init();
+        }
 
-    void setPosition(float pos_)
-    {
-        static_cast<Impl_T*>(this)->_setPosition(pos_);
-    }
+        void _update(void)
+        {
+            static_cast<Impl_T*>(this)->__update();
+        }
 
-    float getPosition(void)
-    {
-        return static_cast<Impl_T*>(this)->_getPosition();
-    }
+        void setPosition(float pos_)
+        {
+            static_cast<Impl_T*>(this)->_setPosition(pos_);
+        }
 
-    void setSpeed(float speed_)
-    {
-        static_cast<Impl_T*>(this)->_setSpeed(speed_);
-    }
+        float getPosition(void)
+        {
+            return static_cast<Impl_T*>(this)->_getPosition();
+        }
 
-    float getSpeed(void)
-    {
-        return static_cast<Impl_T*>(this)->_getSpeed();
-    }
+        void setSpeed(float speed_)
+        {
+            static_cast<Impl_T*>(this)->_setSpeed(speed_);
+        }
 
-    void setMaxSpeed(float max_speed_)
-    {
-        static_cast<Impl_T*>(this)->_setMaxSpeed(max_speed_);
-    }
+        float getSpeed(void)
+        {
+            return static_cast<Impl_T*>(this)->_getSpeed();
+        }
 
-    void setJointLimit(float min_, float max_)
-    {
-        static_cast<Impl_T*>(this)->_setJointLimit(min_, max_);
-    }
+        void setMaxSpeed(float max_speed_)
+        {
+            static_cast<Impl_T*>(this)->_setMaxSpeed(max_speed_);
+        }
 
-    void setReversed(bool reversed_)
-    {
-        static_cast<Impl_T*>(this)->_setReversed(reversed_);
-    }
-};
+        void setJointLimit(std::optional<float> min_, std::optional<float> max_)
+        {
+            static_cast<Impl_T*>(this)->_setJointLimit(min_, max_);
+        }
+
+        void setReversed(bool reversed_)
+        {
+            static_cast<Impl_T*>(this)->_setReversed(reversed_);
+        }
+    };
+
+}  // namespace Actuators
 
 #endif  // ACTUATOR_HPP
